Allow create/update projects with description into Gerrit
Allow the ability for jeepyb to create projects using the
description provided in the YAML file.
Also check the project descriptions for existing projects
and if they differ, update it properly.

Depends on: https://review.openstack.org/#/c/109957/
It needs updates on gerritlib project to properly send
descriptions to Gerrit, and retrieve project
descriptions from it.

Change-Id: Ib92248a3fbbded91d4a250d978e2a50a7bca4e32
Closes-Bug: #1349361